% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_docs.R
\name{Beyer2020}
\alias{Beyer2020}
\title{Documentation for the Beyer2020 dataset}
\description{
This dataset covers the last 120k years, at intervals of 1/2 k years, and a
resolution of 0.5 degrees in latitude and longitude.
}
\details{
If you use this dataset, make sure to cite the original publication:

Beyer, R.M.,
Krapp, M. & Manica, A. High-resolution terrestrial climate, bioclimate and
vegetation for the last 120,000 years. Sci Data 7, 236 (2020).
\doi{10.1038/s41597-020-0552-1}

The version included in \code{pastclim} has the ice sheets masked, as well as
internal seas (Black and Caspian Sea) removed. The latter are based on:

\url{https://www.marineregions.org/gazetteer.php?p=details&id=4278}

\url{https://www.marineregions.org/gazetteer.php?p=details&id=4282}

As there is no reconstruction of their depth through time, modern outlines
were used for all time steps.

Also, for bio15, the coefficient of variation was computed after adding one
to monthly estimates, and it was multiplied by 100 following
\url{https://pubs.usgs.gov/ds/691/ds691.pdf}

Changelog

v1.1.0 Added monthly variables. Files can be downloaded from:
\url{https://zenodo.org/deposit/7062281}

v1.0.0 Remove ice sheets and internal seas, and use correct formula for bio15.
Files can be downloaded from:
\doi{10.6084/m9.figshare.19723405.v1}
}
