% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biome4_classes}
\alias{biome4_classes}
\title{BIOME4 classes.}
\format{
A data.frame with multiple columns to describe.
}
\usage{
biome4_classes
}
\description{
A data.frame defining the details of each class
}
\keyword{datasets}
