% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_datasets.R
\name{check_available_dataset}
\alias{check_available_dataset}
\title{Check if dataset is available.}
\usage{
check_available_dataset(dataset)
}
\arguments{
\item{dataset}{string defining dataset}
}
\description{
Internal getter function
}
\keyword{internal}
