% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars_for_dataset.R
\name{check_available_variable}
\alias{check_available_variable}
\title{Check if var is available for this dataset.}
\usage{
check_available_variable(variable, dataset)
}
\arguments{
\item{variable}{a vector of names of the variables of interest}

\item{dataset}{dataset of interest}
}
\description{
Internal getter function
}
\keyword{internal}
