% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coords_names.R
\name{check_coords_names}
\alias{check_coords_names}
\title{Check that we have a valid pair of coordinate names}
\usage{
check_coords_names(data, coords)
}
\arguments{
\item{data}{a data.frame containing the locations.}

\item{coords}{a vector of length two giving the names of the "x" and "y"
coordinates, of points is a data.frame and does not use standard names.}
}
\value{
A vector of length 2 with the valid names, in the correct order
}
\description{
This internal function checks that coords (as passed to functions) is a valid
set of names, or, if NULL, that we have standard variable names in data
}
\keyword{internal}
