% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataset_path.R
\name{check_dataset_path}
\alias{check_dataset_path}
\title{Check dataset and path_to_nc params}
\usage{
check_dataset_path(dataset, path_to_nc)
}
\arguments{
\item{dataset}{string defining the dataset to use. If set to "custom",
then a single nc file is used from "path_to_nc".}

\item{path_to_nc}{the path to the custom nc file containing the palaeoclimate
reconstructions. All the variables of interest need to be included in
this file.}
}
\value{
TRUE if both dataset and path are valid.
}
\description{
Check that the dataset and path_to_nc parameters are valid.
}
