% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_times.R
\name{check_time_vars}
\alias{check_time_vars}
\title{Check multiple time variables}
\usage{
check_time_vars(time_bp, time_ce, allow_null = TRUE)
}
\arguments{
\item{time_bp}{times in bp}

\item{time_ce}{times in ce}

\item{allow_null}{boolean whether both can be NULL}
}
\value{
times in bp
}
\description{
Check that we only have one set of times
}
\keyword{internal}
