% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downloaded_datasets.R
\name{check_var_downloaded}
\alias{check_var_downloaded}
\title{Internal functions to check whether we have downloaded a given variable for a dataset}
\usage{
check_var_downloaded(variable, dataset, path_to_nc = NULL)
}
\arguments{
\item{variable}{a vector of names of the variables of interest}

\item{dataset}{dataset of interest}

\item{path_to_nc}{path to the netcdf datasets}
}
\description{
Internal functions to check whether we have downloaded a given variable for a dataset
}
\keyword{internal}
