% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_var_in_nc.R
\name{check_var_in_nc}
\alias{check_var_in_nc}
\title{Check whether variables exist in a netcdf file}
\usage{
check_var_in_nc(bio_variables, path_to_nc)
}
\arguments{
\item{bio_variables}{vector of names of variables to be extracted}

\item{path_to_nc}{the path to the custom nc file containing the palaeoclimate
reconstructions.}
}
\value{
TRUE if the variable exists
}
\description{
Internal function to test a custom nc file.
}
\keyword{internal}
