% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_docs.R
\name{CHELSA_trace21k_1.0}
\alias{CHELSA_trace21k_1.0}
\alias{CHELSA_trace21k_1.0_0.5m_vsi}
\title{Documentation for \emph{CHELSA-TracCE21k}}
\description{
CHELSA-TraCE21k data provides monthly climate data for temperature and
precipitation at 30 arc-sec spatial resolution in 100-year time steps for the
last 21,000 years. Palaeo-orography at high spatial resolution and at each
time step is created by combining high resolution information on glacial
cover from current and Last Glacial Maximum (LGM) glacier databases with the
interpolation of a dynamic ice sheet model (ICE6G) and a coupling to mean
annual temperatures from CCSM3-TraCE21k. Based on the reconstructed
palaeo-orography, mean annual temperature and precipitation was downscaled
using the CHELSA V1.2 algorithm.
}
\details{
More details on the dataset are available on its dedicated
\href{https://chelsa-climate.org/chelsa-trace21k/}{website}.

An alternative to downloading very large files is to use virtual rasters.
Simply append "_vis" to the name of the dataset of interest
(\emph{CHELSA_trace21k_1.0_0.5m_vsi}). This is the recommended approach, and it is
currently the only available version of the dataset.

IMPORTANT: If you use this dataset, make sure to cite the original
publication:

Karger, D.N., Nobis, M.P., Normand, S., Graham, C.H., Zimmermann, N. (2023)
CHELSA-TraCE21k – High resolution (1 km) downscaled transient temperature and
precipitation data since the Last Glacial Maximum. Climate of the Past.
\doi{10.5194/cp-2021-30}
}
\note{
There is a missing time slice at -130 years BP for \code{temperature_min_07}
and \code{temperature_max_07} on the CHELSA server. That timestep is replaced with
-120 years BP in \code{pastclim} for those two variables.
}
