% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_path.R
\name{clean_data_path}
\alias{clean_data_path}
\title{Clean the data path}
\usage{
clean_data_path(ask = TRUE)
}
\arguments{
\item{ask}{boolean on whether the user should be asked before deleting}
}
\value{
TRUE if files are deleted successfully
}
\description{
This function deletes old reconstructions that have been superseded in the
data_path. It assumes that the only files in data_path are part of \code{pastclim}
(i.e. there are no custom datasets stored in that directory).
}
