% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data_path.R
\name{copy_example_data}
\alias{copy_example_data}
\title{Internal function to copy the example dataset when a new data path is set}
\usage{
copy_example_data()
}
\value{
TRUE if the data were copied successfully
}
\description{
Copy example dataset
}
\keyword{internal}
