% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclim_vars.R
\name{.cv}
\alias{.cv}
\title{Coefficient of variables (expressed as percentage)}
\usage{
.cv(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
the cv
}
\description{
R function to compute the coefficient of variation
(expressed as a percentage). If there is only a single value, stats::sd = NA.
However, one could argue that cv =0; and NA may break the code that
receives it. The function returns 0 if the mean is close to zero.
}
\details{
This is ODD: abs to avoid very small (or zero) mean with e.g. -5:5
}
\keyword{internal}
