% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_chelsa_trace21k.R
\name{download_chelsa_trace21k}
\alias{download_chelsa_trace21k}
\title{Download the CHELSA trace21k}
\usage{
download_chelsa_trace21k(dataset, bio_var, filename = NULL, time_bp = NULL)
}
\arguments{
\item{dataset}{the name of the dataset}

\item{bio_var}{the variable name}

\item{filename}{the filename as stored in the \code{data_path} of \code{pastclim} (includes the full data path)}

\item{time_bp}{the time steps for which the dataset should be built (NULL for all time steps)}
}
\value{
TRUE if the requested CHELSA variable was downloaded successfully
}
\description{
This function downloads annual and monthly variables from the CHELSA trace v1.0 dataset.
}
\details{
As this dataset is huge, we should not download all files in most situations. For this
reason, time_bp has to be set for downloading (but it is allowed for virtual datasets)
}
\keyword{internal}
