% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download paeloclimate reconstructions.}
\usage{
download_dataset(dataset, bio_variables = NULL)
}
\arguments{
\item{dataset}{string defining dataset to be downloaded (currently only
"Beyer" is available)}

\item{bio_variables}{one or more variable names to be downloaded. If left
to NULL, all variables available for this dataset will be downloaded}
}
\description{
This function downloads paleoclimate reconstructions
}
