% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_paleoclim.R
\name{download_paleoclim}
\alias{download_paleoclim}
\title{Download the paleoclim time series.}
\usage{
download_paleoclim(dataset, bio_var, filename = NULL)
}
\arguments{
\item{dataset}{the name of the dataset}

\item{bio_var}{the variable name}

\item{filename}{(NOT USED FOR THIS FUNCTION: the data come as zip of all bio
variables, so we generate multiple files, not a single one)}
}
\value{
TRUE if the requested paleoclim variable was downloaded successfully
}
\description{
This function downloads annual and monthly variables from the CHELSA v2.1 dataset.
}
\keyword{internal}
