% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_worldclim.R
\name{download_worldclim}
\alias{download_worldclim}
\title{Download a WorldClim modern observations.}
\usage{
download_worldclim(var, res, path = NULL, version = "2.1", ...)
}
\arguments{
\item{var}{character Valid variables names are "tmin", "tmax", "tavg",
"prec", "wind", "vapr", and "bio".}

\item{res}{numeric Valid resolutions are 10, 5, 2.5,
and 0.5 (minutes of a degree)}

\item{path}{character. Path where to download the data to, If left NULL, the data
will be downloaded from the directory returned by \code{\link[=get_data_path]{get_data_path()}}}

\item{version}{character or numeric. WorldClim version number. Only "2.1" supported at the moment}

\item{...}{additional arguments passed to download.file}
}
\value{
TRUE if the requested WorldClim variable was downloaded successfully
}
\description{
This function downloads annual and monthly variables from the WorldClim 2.1 dataset.
These variables are saved in a format that can be read by
\code{\link[=load_worldclim]{load_worldclim()}}, and easily used for delta downscaling palaeoclimate
observations
}
\details{
This function uses \code{\link[geodata:worldclim]{geodata::worldclim_global()}} to download the data, and
then converts them into a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} in a format compatible
with \code{pastclim}
}
\keyword{internal}
