% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_worldclim_present.R
\name{download_worldclim_present}
\alias{download_worldclim_present}
\title{Download a WorldClim modern observations.}
\usage{
download_worldclim_present(dataset, bio_var, filename)
}
\arguments{
\item{dataset}{the name of the dataset}

\item{bio_var}{the variable name}

\item{filename}{the file name (full path) of the file to be saved}
}
\value{
TRUE if the requested WorldClim variable was downloaded successfully
}
\description{
This function downloads annual and monthly variables from the WorldClim 2.1
dataset.
}
\keyword{internal}
