% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_docs.R
\name{Example}
\alias{Example}
\title{Documentation for the Example dataset}
\description{
This dataset is a subset of Beyer2020, used for the vignette of pastclim. Do
not use this dataset for any real work, as it might not reflect the most
up-to-date version of Beyer2020.
}
