% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames_chelsa_future.R
\name{filenames_chelsa_future}
\alias{filenames_chelsa_future}
\title{Generate file names to download the chelsa future dataset}
\usage{
filenames_chelsa_future(dataset, bio_var)
}
\arguments{
\item{dataset}{the name of the dataset of interest}

\item{bio_var}{the variable of interest}
}
\value{
a vector of times, one per band
}
\description{
This function creates a vector of paths needed to download the CHELSA future
dataset
}
\keyword{internal}
