% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames_chelsa_trace21k.R
\name{filenames_chelsa_trace21k}
\alias{filenames_chelsa_trace21k}
\title{Generate file names to download the chelsa trace 21k}
\usage{
filenames_chelsa_trace21k(dataset, bio_var)
}
\arguments{
\item{dataset}{the name of the dataset of interest (currently unused)}

\item{bio_var}{the variable of interest}
}
\value{
a vector of times, one per band
}
\description{
This function creates a vector of paths needed to download the CHELSA trace21k
}
\keyword{internal}
