% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames_paleoclim.R
\name{filenames_paleoclim}
\alias{filenames_paleoclim}
\title{Generate file names to download the chelsa present dataset}
\usage{
filenames_paleoclim(dataset, bio_var, version = NULL)
}
\arguments{
\item{dataset}{the name of the dataset of interest (currently unused)}

\item{bio_var}{the variable of interest}

\item{version}{the version of the dataset (currently unused)}
}
\value{
a vector of times, one per band
}
\description{
This function creates a vector of paths needed to download the CHELSA present
dataset. Possible names are "paleoclim_1.0_10m", "paleoclim_1.0_5m",
"paleoclim_1.0_2.5m"
}
\keyword{internal}
