% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames_worldclim_present.R
\name{filenames_worldclim_present}
\alias{filenames_worldclim_present}
\title{Generate file names to download the WorldClim present dataset}
\usage{
filenames_worldclim_present(dataset, bio_var)
}
\arguments{
\item{dataset}{the name of the dataset of interest}

\item{bio_var}{the variable of interest}
}
\value{
a vector of times, one per band
}
\description{
This function creates a vector of paths needed to download the WorldClim
present dataset
}
\keyword{internal}
