% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_datasets.R
\name{get_available_datasets}
\alias{get_available_datasets}
\title{Get the available datasets.}
\usage{
get_available_datasets()
}
\value{
a character vector of the available datasets
}
\description{
List the datasets available in pastclim, which can be passed to functions
in \code{pastclim} as values for the parameter \code{dataset}. Most functions can also be
used on custom datasets by setting \code{dataset="custom"}
}
\details{
This function provides a user-friendly list, summarising the many datasets
available from WorldClim. A comprehensive list of all available datasets
can be obtained with \link{list_available_datasets}.
}
