% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biome_classes.R
\name{get_biome_classes}
\alias{get_biome_classes}
\title{Get the biome classes for a dataset.}
\usage{
get_biome_classes(dataset)
}
\arguments{
\item{dataset}{string defining dataset to be downloaded (a list of possible
values can be obtained with \code{get_available_datasets}). This function
will not work on custom datasets.}
}
\value{
a data.frame with columns id and category.
}
\description{
Get a full list of biomes and how their id as coded in the biome variable
for a given dataset.
}
