% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_info.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{Get the information about a dataset}
\usage{
get_dataset_info(dataset)
}
\arguments{
\item{dataset}{A dataset in pastclim}
}
\description{
This function provides full information about a given dataset. A full
list of datasets available in pastclim can be obtained with
\code{get_available_datasets}
}
\keyword{internal}
