% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downloaded_datasets.R
\name{get_downloaded_datasets}
\alias{get_downloaded_datasets}
\title{Get the variables downloaded for each dataset.}
\usage{
get_downloaded_datasets(path_to_nc = NULL)
}
\arguments{
\item{path_to_nc}{path to the netcdf datasets}
}
\description{
List the downloaded variable for each dataset.
}
