% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mis_time_steps.R
\name{get_mis_time_steps}
\alias{get_mis_time_steps}
\title{Get time steps for a given MIS}
\usage{
get_mis_time_steps(mis, dataset = NULL, path_to_nc = NULL)
}
\arguments{
\item{mis}{string giving the mis; it must use the same spelling as used in
/code{mis_boundaries}}

\item{dataset}{string defining dataset to be downloaded (currently only
"Beyer2020" or "Krapp2021" are available)}

\item{path_to_nc}{the path to the directory containing the downloaded
reconstructions. Leave it unset if you are using the companion
`pastclimData` to store datasets.}
}
\description{
Get the time steps available in a given dataset for a MIS.
}
