% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pastclimdata_path.R
\name{get_pastclimdata_path}
\alias{get_pastclimdata_path}
\title{Get the path for pastclimdata.}
\usage{
get_pastclimdata_path()
}
\description{
This function returns the path for pastclimData where reconstructions
are stored.
}
