% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_res.R
\name{get_resolution}
\alias{get_resolution}
\title{Get resolution of a given dataset}
\usage{
get_resolution(dataset, path_to_nc = NULL)
}
\arguments{
\item{dataset}{string defining dataset to be downloaded (a list of possible
values can be obtained with \code{\link[=list_available_datasets]{list_available_datasets()}}). If set to
"custom", then a single nc file is used from "path_to_nc"}

\item{path_to_nc}{the path to the custom nc file containing the palaeoclimate
reconstructions. All the variables of interest need to be included in
this file.}
}
\value{
a vector of resolution in the x and y axes
}
\description{
Get the resolution of a given dataset.
}
