% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_steps.R
\name{get_time_steps}
\alias{get_time_steps}
\title{Get time steps for a given dataset}
\usage{
get_time_steps(dataset = NULL, path_to_nc = NULL)
}
\arguments{
\item{dataset}{string defining dataset to be downloaded (currently only
"Beyer2020", "Krapp2021" and "Example" are available). Leave it unset
if using `path_to_nc`}

\item{path_to_nc}{the path to the directory containing the downloaded
reconstructions. Leave it unset if you are using the companion
`pastclimData` to store datasets.}
}
\description{
Get the time steps available in a given dataset.
}
