% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_for_dataset.R
\name{get_var_meta}
\alias{get_var_meta}
\title{Get the metadata for a variable in a given dataset.}
\usage{
get_var_meta(variable, dataset)
}
\arguments{
\item{variable}{one or more variable names to be downloaded}

\item{dataset}{string defining dataset to be downloaded (a list of possible
values can be obtained with \code{\link[=list_available_datasets]{list_available_datasets()}}). This function
will not work on custom datasets.}
}
\value{
the metadata (including filename) for that variable in that dataset
}
\description{
Internal getter function
}
\keyword{internal}
