% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars_for_dataset.R
\name{get_varname}
\alias{get_varname}
\title{Get a the varname for this variable}
\usage{
get_varname(variable, dataset)
}
\arguments{
\item{variable}{string defining the variable name}

\item{dataset}{string defining dataset to be downloaded}
}
\value{
the name of the variable
}
\description{
Internal function to get the varname for this variable
}
\keyword{internal}
