% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars_for_dataset.R
\name{get_vars_for_dataset}
\alias{get_vars_for_dataset}
\title{Get a list of variables for a given dataset.}
\usage{
get_vars_for_dataset(dataset)
}
\arguments{
\item{dataset}{string defining dataset to be downloaded (a list of possible
values can be obtained with \code{get_available_datasets}). This function
will not work on custom datasets.}
}
\description{
This function lists the variables available for a given dataset. Note that
the spelling and use of capitals in names might differ from the original
publications, as `pastclim` harmonises the names of variables across
different reconstructions.
}
