% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_info.R
\name{help_console}
\alias{help_console}
\title{Print help to console}
\usage{
help_console(
  topic,
  format = c("text", "html", "latex"),
  lines = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{topic}{The topic of the help}

\item{format}{how the output should be formatted}

\item{lines}{which lines should be printed}

\item{before}{string to be printed before the output}

\item{after}{string to be printed after the output}
}
\value{
text of the help file
}
\description{
This function prints a help file to console. It is based on a function
published on R-bloggers:
from https://www.r-bloggers.com/2013/06/printing-r-help-files-in-the-console-or-in-knitr-documents/
}
\keyword{internal}
