% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{koeppen_classes}
\alias{koeppen_classes}
\title{Koeppen-Geiger classes.}
\format{
A data.frame with multiple columns to describe.
}
\usage{
koeppen_classes
}
\description{
A data.frame defining the details of each class
}
\keyword{datasets}
