% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_available_datasets.R
\name{list_available_datasets}
\alias{list_available_datasets}
\title{List all the available datasets.}
\usage{
list_available_datasets()
}
\value{
a character vector of the available datasets
}
\description{
List the datasets available in pastclim. The list is comprehensive,
and includes all combinations of models and future scenarios for WorldClim.
For a more user-friendly list, use \code{\link[=get_available_datasets]{get_available_datasets()}}. Most functions can also be
used on custom datasets by setting \code{dataset="custom"}
}
