% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset_list.R
\name{load_dataset_list}
\alias{load_dataset_list}
\title{Load the dataset list}
\usage{
load_dataset_list(on_cran = FALSE)
}
\arguments{
\item{on_cran}{boolean to make this function run on ci tests using tempdir}
}
\value{
the dataset list
}
\description{
This function returns a dataframe with the details for each variable
available in every dataset. It defaults to the copy stored within the
package, but it checks in case there is an udpated version stored as
'data_list.csv' in 
`tools::R_user_dir("pastclim","config")`. If the latter is present, the last
column, named 'dataset_list_v', provides the version of this table, and the
most advanced table is used.
}
\keyword{internal}
