% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_worldclim.R
\name{load_worldclim}
\alias{load_worldclim}
\title{Load a variable from WorldClim modern observations.}
\usage{
load_worldclim(var, res, path = NULL, version = "2.1", ...)
}
\arguments{
\item{var}{character Valid variables names are "tmin", "tmax", "tavg",
"prec", "wind", "vapr", and "bio".}

\item{res}{numeric Valid resolutions are 10, 5, 2.5,
and 0.5 (minutes of a degree)}

\item{path}{character. Path where the dataset is stored. If left NULL, the data
will be downloaded from the directory returned by \code{\link[=get_data_path]{get_data_path()}}}

\item{version}{character or numeric. WorldClim version number. Only "2.1" supported at the moment}
}
\value{
a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} with the requested WorldClim variable
}
\description{
This function loads previously downloaded annual and monthly variables
from the WorldClim 2.1 dataset. To save the variables in a compatible
format, use \code{\link[=download_worldclim]{download_worldclim()}}.
}
\details{
The function assumes that the file name is in the format "wc\emph{version}\emph{\emph{res}m}\emph{var}.nc".
For example, average temperature downloaded at 10 minutes resolution will be
"\emph{wc21_10m_tavg.nc}".
}
\keyword{internal}
