% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mis_boundaries}
\alias{mis_boundaries}
\title{Time boundaries of marine isotope stages (MIS).}
\format{
A data frame with 24 rows and 2 variables:
\describe{
  \item{mis}{the stage, a string}
  \item{start}{the start of a given MIS, in kya}
  \item{end}{the start of a given MIS, in kya}
}
}
\usage{
mis_boundaries
}
\description{
A dataset containing the beginning and end of MIS.
}
\keyword{datasets}
