% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_docs.R
\name{paleoclim_1.0}
\alias{paleoclim_1.0}
\alias{paleoclim_1.0_10m}
\title{Documentation for paleoclim}
\description{
Stub description for the paleoclim dataset
}
\details{
If you use this dataset, make sure to cite the original publication
for the paleclim dataset:

Brown, Hill, Dolan, Carnaval, Haywood (2018) PaleoClim, high spatial
resolution paleoclimate surfaces for global land areas.
Nature – Scientific Data. 5:180254
}
