% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastclim_rast.R
\name{pastclim_rast}
\alias{pastclim_rast}
\title{Read a raster for pastclim}
\usage{
pastclim_rast(
  x,
  bio_var_orig,
  bio_var_pastclim,
  var_longname = NULL,
  var_units = NULL
)
}
\arguments{
\item{x}{filename of the raster}

\item{bio_var_orig}{the variable name as present in the file}

\item{bio_var_pastclim}{the variable name as used by pastclim (thus allowing
us to rename the variable)}
}
\value{
a \code{\link[terra:SpatRaster]{terra::SpatRaster}}
}
\description{
This function is a wrapper around \code{\link[terra:rast]{terra::rast()}}, with additional logic to
correctly import time for vrt datasets (time is stored as custom metadata in
pastclim-generated vrt files)
}
\keyword{internal}
