% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_extent}
\alias{region_extent}
\title{Region extents.}
\format{
A list of vectors giving the extents.
}
\usage{
region_extent
}
\description{
A list of extents for major regions.
}
\keyword{datasets}
