% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_outline}
\alias{region_outline}
\title{Region outlines.}
\format{
\code{\link[sf:sf]{sf::sf}} of outlines.
\describe{
\item{name}{names of regions}
}
}
\usage{
region_outline
}
\description{
An \code{\link[sf:sf]{sf::sf}} object containing outlines for major regions. Outlines that
span the antimeridian have been split into multiple polygons.
}
\keyword{datasets}
