% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_outline_union}
\alias{region_outline_union}
\title{Region outlines unioned.}
\format{
\code{sf} of outlines.
\describe{
  \item{name}{names of regions}
}
}
\usage{
region_outline_union
}
\description{
An \code{sf} object containing outlines for major regions. Each outline is
represented as a single polygon. If you want multiple polygons, use
\link{region_outline}.
}
\keyword{datasets}
