% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data_path.R
\name{set_data_path_for_CRAN}
\alias{set_data_path_for_CRAN}
\title{Set the data path for examples on CRAN}
\usage{
set_data_path_for_CRAN()
}
\value{
None
}
\description{
Users should NOT need this function. It is used to set up a
data path in the temporary directory for examples and tests
to run on CRAN.
}
