% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bp_series.R
\name{time_bp_series}
\alias{time_bp_series}
\title{Convert a time BP to indexes for a series}
\usage{
time_bp_series(time_bp, time_steps)
}
\arguments{
\item{time_bp}{vector of times BP}

\item{time_steps}{time steps for which reconstructions are available}
}
\value{
the indeces for the relevant time steps
}
\description{
Internal function
}
\keyword{internal}
