% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_series.R, R/location_series_old.R
\name{time_series_for_locations}
\alias{time_series_for_locations}
\title{Extract a time series of bioclimatic variables for one or more locations.}
\usage{
time_series_for_locations(...)

time_series_for_locations(...)
}
\arguments{
\item{...}{arguments to be passed to \code{series}}
}
\value{
a data.frame with the climatic variables of interest

a data.frame with the climatic variables of interest
}
\description{
Deprecated version of \code{location_series}

Deprecated version of \code{location_series}
}
