% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_is_valid.R
\name{url_is_valid}
\alias{url_is_valid}
\title{Test whether a URL is valid}
\usage{
url_is_valid(url, verbose = FALSE)
}
\arguments{
\item{url}{the url to test}

\item{verbose}{whether the status code should be outputted as a message}
}
\value{
boolean on whether the file exists
}
\description{
This function check that a URL points to a real file
}
\keyword{internal}
