% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_nc.R
\name{validate_nc}
\alias{validate_nc}
\title{Validate an netcdf file for pastclim}
\usage{
validate_nc(path_to_nc)
}
\arguments{
\item{path_to_nc}{path to the nc file of interest}
}
\value{
TRUE if the file is valid.
}
\description{
This function validates a netcdf file as a potential dataset for \code{pastlim}.
The key checks are:
a) that the dimensions (longitude, latitude and time) have been set correctly.
b) that all variables have the appropriate metadata (longname and units)
}
