% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt_get_times.R
\name{vrt_get_times}
\alias{vrt_get_times}
\title{Get time axis from vrt}
\usage{
vrt_get_times(vrt_path)
}
\arguments{
\item{vrt_path}{path to the XML file defining the vrt dataset}
}
\value{
a vector of times, one per band
}
\description{
This function extract time information for the bands in a vrt. It first checks
that the vrt dataset has a metadata element with key "has_time" set to TRUE. If that
is the case, for each band, it extract the metadata with key "time" and returns
them as numeric (i.e. converting them from character).
Note that an error is returned if there are duplicated time elements for any of the
bands (whilst duplicated elements are valid in the XML schema for VRT, they do
not make sense for the time axis).
}
\keyword{internal}
