% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt_set_meta.R
\name{vrt_set_meta}
\alias{vrt_set_meta}
\title{Set vrt metadata}
\usage{
vrt_set_meta(vrt_path, description, time_vector)
}
\arguments{
\item{vrt_path}{path to the XML file defining the vrt dataset}

\item{description}{a string with the description of the variable in this dataset}

\item{time_vector}{a vector of descriptions (same length as the number of bands)}
}
\value{
TRUE if the file was updated correctly
}
\description{
This function sets the metadata information to a vrt file. It creates some dataset
wide metadata, as well as band specific descriptions and times.
}
\keyword{internal}
