% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bathy_to_spatraster.R
\name{bathy_to_spatraster}
\alias{bathy_to_spatraster}
\title{Cast \code{bathy} to \code{SpatRaster}}
\usage{
bathy_to_spatraster(bathy)
}
\arguments{
\item{bathy}{a \code{\link[marmap:as.bathy]{marmap::bathy}} to convert}
}
\value{
a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} with the relief for the chosen region
}
\description{
This function converts a \code{\link[marmap:as.bathy]{marmap::bathy}} object to
a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}.
}
\keyword{internal}
