% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bathy_to_spatraster.R
\name{bathy_to_spatraster}
\alias{bathy_to_spatraster}
\title{Cast bathy to SpatRaster}
\usage{
bathy_to_spatraster(bathy)
}
\arguments{
\item{bathy}{a \code{marmap::bathy} to convert}
}
\value{
a \code{terra::SpatRaster} with the relief for the chosen region
}
\description{
This function converts a \code{marmap::bathy} object to 
a \code{terra::SpatRaster}.
}
\keyword{internal}
