% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_var_downloaded.R
\name{check_var_downloaded}
\alias{check_var_downloaded}
\title{Internal functions to check whether we have downloaded a given variable
for a dataset}
\usage{
check_var_downloaded(variable, dataset)
}
\arguments{
\item{variable}{a vector of names of the variables of interest}

\item{dataset}{dataset of interest}
}
\description{
Internal functions to check whether we have downloaded a given variable
for a dataset
}
\keyword{internal}
