% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_series.R, R/location_slice.R
\name{climate_for_locations}
\alias{climate_for_locations}
\title{Extract a time series of bioclimatic variables for one or more locations.}
\usage{
climate_for_locations(...)

climate_for_locations(...)
}
\arguments{
\item{...}{arguments to be passed to \code{location_slice}}
}
\description{
Deprecated version of \code{location_series}

Deprecated version of \code{location_slice}
}
