% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_slice.R
\name{climate_for_locations}
\alias{climate_for_locations}
\title{Extract local climate for one or more locations for a given time slice.}
\usage{
climate_for_locations(...)
}
\arguments{
\item{...}{arguments to be passed to \code{\link[=location_slice]{location_slice()}}}
}
\value{
a data.frame with the climatic variables of interest
}
\description{
Deprecated version of \code{\link[=location_slice]{location_slice()}}
}
