% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_slice.R
\name{climate_for_time_slice}
\alias{climate_for_time_slice}
\title{Extract a climate slice for a region}
\usage{
climate_for_time_slice(...)
}
\arguments{
\item{...}{arguments to be passed to \code{region_slice}}
}
\value{
a
SpatRaster \code{terra::SpatRaster} object, with
each variable as a layer.
}
\description{
Deprecated version of \code{region_slice}
}
