% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_for_time_slice.R
\name{climate_for_time_slice}
\alias{climate_for_time_slice}
\title{Extract a raster brick for a given time slice}
\usage{
climate_for_time_slice(time_bp, bio_variables, dataset, path_to_nc = NULL)
}
\arguments{
\item{time_bp}{time slice in years before present (negative)}

\item{bio_variables}{vector of names of variables to be extracted}

\item{dataset}{string defining the dataset to use}

\item{path_to_nc}{the path to the directory containing the downloaded resonstructions.
Leave it unset if you are using the companion `pastclimData` to store datasets.}
}
\description{
This function extracts a raster brick for a given time slice.
}
