% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_from_region_slice.R
\name{df_from_region_slice}
\alias{df_from_region_slice}
\title{Extract data frame from a region slice}
\usage{
df_from_region_slice(x, xy = TRUE)
}
\arguments{
\item{x}{climate time slice (i.e. a \code{terra::SpatRaster}) 
generated with \code{region_slice}}

\item{xy}{a boolean whether x and y coordinates should be added 
to the dataframe (default to TRUE)}
}
\description{
Extract the climatic information from a region slice and organise it as
a data frame. This is just a wrapper around \code{terra::as.data.frame}.
}
\details{
To extract a data frame from a region series, see \code{df_from_region_series}.
}
