% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_from_sea.R
\name{distance_from_sea}
\alias{distance_from_sea}
\title{Compute a raster of distances from the sea for each land pixel.}
\usage{
distance_from_sea(time_bp, dataset)
}
\arguments{
\item{time_bp}{time slice in years before present (negative)}

\item{dataset}{string defining dataset to be downloaded (a list of possible
values can be obtained with \code{get_available_datasets}). This function
will not work on custom datasets.}
}
\value{
a \code{terra::SpatRaster} of distances
}
\description{
Get the land mask for a dataset at a given time point, and compute distance
from the sea for each land pixel.
}
