% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_chelsa.R
\name{download_chelsa}
\alias{download_chelsa}
\title{Download the CHELSA modern and future observations.}
\usage{
download_chelsa(dataset, bio_var, filename)
}
\arguments{
\item{dataset}{the name of the dataset}

\item{bio_var}{the variable name}

\item{filename}{the filename as stored in the \code{data_path} of \code{pastclim}
(includes the full data path)}
}
\value{
TRUE if the requested CHELSA variable was downloaded successfully
}
\description{
This function downloads annual and monthly variables from the CHELSA v2.1
dataset.
}
\keyword{internal}
