% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download paeloclimate reconstructions.}
\usage{
download_dataset(dataset, bio_variables = NULL, path_to_nc = NULL)
}
\arguments{
\item{dataset}{string defining dataset to be downloaded (currently only
"Beyer" is available)}

\item{bio_variables}{one or more variable names to be downloaded. If left
to NULL, all variables available for this dataset will be downloaded}

\item{path_to_nc}{directory where the files will be saved. If not set, the
data will be downloaded into the storage package pastclimData; an error
will be returned if this package is not installed.}
}
\description{
This function downloads paleoclimate reconstructions
}
