% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_etopo.R
\name{download_etopo}
\alias{download_etopo}
\title{Download the ETOPO Global relief model}
\usage{
download_etopo(path = NULL, resolution = 60)
}
\arguments{
\item{path}{character. Path where to download the data to. If left NULL, the data
will be downloaded from the directory returned by \code{\link[=get_data_path]{get_data_path()}}}

\item{resolution}{numeric resolution in arcsecs (one of 30, or 60).
Defaults to 60 arcsecs.}
}
\value{
TRUE if the dataset was downloaded successfully
}
\description{
This function downloads the ETOPO2022 global relief model at 30 or 60 arcsecs
resolution. This is a large file (>1Gb), and is worth downloading only
if you are planning to use it repeatedly. If you only need a subregion,
especially at lower resolution, it is worth using \code{\link[=download_etopo_subset]{download_etopo_subset()}},
which fetches a subset of the dataset on the fly from the NOAA server.
}
\keyword{internal}
