% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_worldclim_future.R
\name{download_worldclim_future}
\alias{download_worldclim_future}
\title{Download a WorldClim future predictions.}
\usage{
download_worldclim_future(dataset, bio_var, filename)
}
\arguments{
\item{dataset}{the name of the dataset}

\item{bio_var}{the variable name}

\item{filename}{the file name (full path) of the file to be saved}
}
\value{
TRUE if the requested WorldClim variable was downloaded successfully
}
\description{
This function downloads annual and monthly variables from the WorldClim 2.1
predictions for the future.
}
\keyword{internal}
