% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_datasets.R
\name{get_available_datasets}
\alias{get_available_datasets}
\title{Get the available datasets.}
\usage{
get_available_datasets()
}
\value{
a character vector of the available datasets
}
\description{
List the datasets available in pastclim. Most functions can also be
used on custom datasets by setting `dataset="custom"`
}
