% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biome_classes.R
\name{get_biome_classes}
\alias{get_biome_classes}
\title{Get the biome classes for a dataset.}
\usage{
get_biome_classes(dataset, path_to_nc = NULL)
}
\arguments{
\item{dataset}{string defining dataset to be downloaded (currently only
"Beyer2020" or "Krapp2021" are available)}

\item{path_to_nc}{the path to the directory containing the downloaded resonstructions.
Leave it unset if you are using the companion `pastclimData` to store datasets.}
}
\description{
Get a full list of biomes and how their id as coded in the biome variable
for a given dataset.
}
