% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data_path.R
\name{get_data_path}
\alias{get_data_path}
\title{Get the data path where climate reconstructions are stored}
\usage{
get_data_path()
}
\value{
the data path
}
\description{
This function returns the path where climate reconstructions will be stored.
When `pastclim` is installed, the data path defaults to the directory
returned by `tools::R_user_dir("pastclim","data")` (i.e. the data directory
for the package in R>=4.0). The data path can be changed with 
\code{set_data_path}. Once changed, `pastclim` will remember the new data
path in the future.
}
\details{
The data path is stored in a file named "pastclim_data.txt", which
is found in the directory returned by 
`tools::R_user_dir("pastclim","config")` (i.e.
the default configuration directory for the package as set in R >= 4.0).
}
