% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downloaded_datasets.R
\name{get_downloaded_datasets}
\alias{get_downloaded_datasets}
\title{Get the variables downloaded for each dataset.}
\usage{
get_downloaded_datasets(data_path = NULL)
}
\arguments{
\item{data_path}{leave it to NULL to use the default datapath}
}
\value{
a list of variable names per dataset.
}
\description{
List the downloaded variable for each dataset.
}
