% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_meta.R
\name{get_var_meta}
\alias{get_var_meta}
\title{Get the file details for a variable and dataset.}
\usage{
get_var_meta(variable, dataset)
}
\arguments{
\item{variable}{one or more variable names to be downloaded}

\item{dataset}{string defining dataset to be downloaded (a list of possible
values can be obtained with \code{\link[=list_available_datasets]{list_available_datasets()}}). This function
will not work on custom datasets.}
}
\value{
the filename for that variable and dataset
}
\description{
Internal getter function
}
