% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_for_dataset.R
\name{get_file_for_dataset}
\alias{get_file_for_dataset}
\title{Get the file details for a variable and dataset.}
\usage{
get_file_for_dataset(variable, dataset)
}
\arguments{
\item{variable}{one or more variable names to be downloaded}

\item{dataset}{string defining dataset to be downloaded (a list of possible
values can be obtained with \code{\link[=get_available_datasets]{get_available_datasets()}}). This function
will not work on custom datasets.}
}
\value{
the filename for that variable and dataset
}
\description{
Internal getter function
}
