% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ice_mask.R
\name{get_ice_mask}
\alias{get_ice_mask}
\title{Get the ice mask for a dataset.}
\usage{
get_ice_mask(time_bp, dataset, path_to_nc = NULL)
}
\arguments{
\item{time_bp}{time slice in years before present (negative)}

\item{dataset}{string defining dataset to be downloaded (currently only
"Beyer2020" or "Krapp2021" are available). Note that this function will not work
on an custom dataset.}

\item{path_to_nc}{the path to the directory containing the downloaded resonstructions.
Leave it unset if you are using the companion `pastclimData` to store datasets.}
}
\description{
Get the ice mask for a dataset at a given timepoint.
}
