% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ice_mask.R
\name{get_ice_mask}
\alias{get_ice_mask}
\title{Get the ice mask for a dataset.}
\usage{
get_ice_mask(time_bp, dataset)
}
\arguments{
\item{time_bp}{time slice in years before present (negative)}

\item{dataset}{string defining dataset to be downloaded (a list of possible
values can be obtained with \code{get_available_datasets}). This function
will not work on custom datasets.}
}
\description{
Get the ice mask for a dataset at a given time point.
}
