% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_land_mask.R
\name{get_land_mask}
\alias{get_land_mask}
\title{Get the land mask for a dataset.}
\usage{
get_land_mask(time_bp, dataset, path_to_nc = NULL)
}
\arguments{
\item{time_bp}{time slice in years before present (negative)}

\item{dataset}{string defining dataset to be downloaded (currently only
"Beyer2020" or "Krapp2021" are available). Note that this function will
not work on an custom dataset.}

\item{path_to_nc}{the path to the directory containing the downloaded
reconstructions. Leave it unset if you are using the companion
`pastclimData` to store datasets.}
}
\description{
Get the land mask for a dataset at a given timepoint.
}
