% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_land_mask.R
\name{get_land_mask}
\alias{get_land_mask}
\title{Get the land mask for a dataset.}
\usage{
get_land_mask(time_bp, dataset)
}
\arguments{
\item{time_bp}{time slice in years before present (negative)}

\item{dataset}{string defining dataset to be downloaded (a list of possible
values can be obtained with \code{get_available_datasets}). This function
will not work on custom datasets.}
}
\value{
a binary \code{terra::SpatRaster} with the land mask as 1s
}
\description{
Get the land mask for a dataset at a given time point.
}
