% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mis_time_steps.R
\name{get_mis_time_steps}
\alias{get_mis_time_steps}
\title{Get time steps for a given MIS}
\usage{
get_mis_time_steps(mis, dataset, path_to_nc = NULL)
}
\arguments{
\item{mis}{string giving the mis; it must use the same spelling as used in
\link{mis_boundaries}}

\item{dataset}{string defining dataset to be downloaded (a list of possible
values can be obtained with \code{\link[=list_available_datasets]{list_available_datasets()}}). If set to
"custom", then a single nc file is used from "path_to_nc"}

\item{path_to_nc}{the path to the custom nc file containing the palaeoclimate
reconstructions. All the variables of interest need to be included in
this file.}
}
\value{
a vector of time steps
}
\description{
Get the time steps available in a given dataset for a MIS.
}
