% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_steps.R
\name{get_time_steps}
\alias{get_time_steps}
\title{Get time steps for a given dataset}
\usage{
get_time_steps(dataset, path_to_nc = NULL)
}
\arguments{
\item{dataset}{string defining dataset to be downloaded (a list of possible
values can be obtained with \code{get_available_datasets}). If set to
"custom", then a single nc file is used from "path_to_nc"}

\item{path_to_nc}{the path to the custom nc file containing the palaeoclimate
reconstructions. All the variables of interest need to be included in
this file.}
}
\description{
Get the time steps (in time_bp) available in a given dataset.
}
