% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars_for_dataset.R
\name{get_vars_for_dataset}
\alias{get_vars_for_dataset}
\title{Get a list of variables for a given dataset.}
\usage{
get_vars_for_dataset(
  dataset,
  path_to_nc = NULL,
  details = FALSE,
  annual = TRUE,
  monthly = FALSE
)
}
\arguments{
\item{dataset}{string defining dataset to be downloaded (a list of possible
values can be obtained with \code{\link[=list_available_datasets]{list_available_datasets()}}).}

\item{path_to_nc}{the path to the custom nc file containing the palaeoclimate
reconstructions. If a custom nc file is given, 'details', 'annual' and 'monthly'
are ignored}

\item{details}{boolean determining whether the output should include information
including long names of variables and their units.}

\item{annual}{boolean to show annual variables}

\item{monthly}{boolean to show monthly variables}
}
\value{
a vector of variable names
}
\description{
This function lists the variables available for a given dataset. Note that
the spelling and use of capitals in names might differ from the original
publications, as \code{pastclim} harmonises the names of variables across
different reconstructions.
}
