% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idw_interp.R
\name{idw_interp}
\alias{idw_interp}
\title{Interpolate x to match mask y}
\usage{
idw_interp(x, y, ...)
}
\arguments{
\item{x}{the \code{terra::SpatRaster} of the variable of interest}

\item{y}{the \code{terra::SpatRaster} the reference mask defining which
cells should have values}

\item{...}{paramters to be passed to \code{gstat::gstat}}
}
\description{
Fill in x to match cells available in y, using iwd (inverse-weighted
distance) interpolation. Interpolation is fitted using \code{gstat::gstat};
the default parameters for \code{gstat::gstat} are nmax=7 and idp = .5, but
can be changed by providing arguments to this function (which will be passed
to \code{gstat::gstat}). See \code{gstat::gstat} for details on the available
parameters and their meaning.
}
\keyword{internal}
