% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_docs.R
\name{Krapp2021}
\alias{Krapp2021}
\title{Documentation for the Krapp2021 dataset}
\description{
This dataset covers the last 800k years, at intervals of 1k years, and a
resolution of 0.5 degrees in latitude and longitude.
}
\details{
The units of several variables have been changed to match what is used in
WorldClim.

IMPORTANT: If you use this dataset, make sure to cite the original
publication:

Krapp, M., Beyer, R.M., Edmundson, S.L. et al. A statistics-based
reconstruction of high-resolution global terrestrial climate for the last
800,000 years. Sci Data 8, 228 (2021). \doi{10.1038/s41597-021-01009-3}

The version included in \code{pastclim} has the ice sheets masked.

Note that, for bio15, we use the corrected version, which follows
\url{https://pubs.usgs.gov/ds/691/ds691.pdf}

Changelog

v1.4.0 Change units to match WorldClim. Fix variable duplication found on
earlier versions of the dataset. \url{https://zenodo.org/records/8415273}

v1.1.0 Added monthly variables. Files can be downloaded from:
\url{https://zenodo.org/record/7065055}

v1.0.0 Remove ice sheets and use correct formula for bio15. Files can be
downloaded from: \doi{10.6084/m9.figshare.19733680.v1}
}
