% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_for_location.R
\name{location_slice}
\alias{location_slice}
\title{Extract local climate for one or more locations for a given time slice.}
\usage{
location_slice(
  x,
  time_bp,
  bio_variables,
  dataset,
  path_to_nc = NULL,
  nn_interpol = TRUE
)
}
\arguments{
\item{x}{a 2 column matrix (with columns `longitude`, ranging
-180 to 180, and `latitude`, from -90 to 90), or a vector of cell numbers.}

\item{time_bp}{vector of ages, in years before present (negative).}

\item{bio_variables}{vector of names of variables to be extracted.}

\item{dataset}{string defining the dataset to use (one of Beyer2020,
Krapp2021, Example or custom).}

\item{path_to_nc}{the path to the custom nc file containing the paleoclimate
reconstructions. All the variables of interest need to be included in this file.}

\item{nn_interpol}{boolean determining whether nearest neighbour
interpolation is used to estimate climate for cells that lack such
information (i.e. they are under water or ice). Interpolation is only
performed from the first ring of nearest neighbours; if climate is not
available, NA will be returned for that location. Defaults to TRUE.}
}
\description{
This function extract local climate for a set of locations
at the appropriate times (selecting the closest time slice available for the
specific date associated with each location).
}
