% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_binary_mask.R
\name{make_binary_mask}
\alias{make_binary_mask}
\title{Create a binary mask}
\usage{
make_binary_mask(x)
}
\arguments{
\item{x}{a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}}
}
\value{
a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} with 0s and 1s
}
\description{
Create a binary mask from a raster: NAs are converted to 0s, and
any other value to 1.
}
\keyword{internal}
