% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_land_mask.R
\name{make_land_mask}
\alias{make_land_mask}
\title{Create a land mask}
\usage{
make_land_mask(topo_rast, time_bp, sea_level = NULL)
}
\arguments{
\item{topo_rast}{a \code{terra::SpatRaster} with topography}

\item{time_bp}{the time of interest}

\item{sea_level}{sea level at the time of interest (if left to NULL, this is
computed using Spratt 2016)}
}
\value{
a \code{terra::SpatRaster} of 0s (sea) and 1s (land), where the
layers are different times
}
\description{
Create a land mask for a given time step. The land mask is based on the simple
logic of moving the ocean up and down given the current topography/bathymetry.
Note that this approach ignores any rebound due to changing mass and distribution of
icesheets. The landmask will show internal lakes/seas as land, as their level
is unrelated to the general sea level. If you have specific reconstructions
of internal lakes (or want to simply reuse their current extents), you will
have to add them onto the masks generated by this function.
}
\keyword{internal}
