% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_slice_from_region_series_fast.R
\name{mode}
\alias{mode}
\title{Mode}
\usage{
mode(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the mode
}
\description{
Find the mode of vector x (note that, if multiple values have the same
frequency, this function simply picks the first occurring one)
}
\keyword{internal}
