% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastclim-package.R
\docType{package}
\name{pastclim-package}
\alias{pastclim}
\alias{pastclim-package}
\title{pastclim: Manipulate Time Series of Climate Reconstructions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Methods to easily extract and manipulate climate reconstructions for ecological and anthropological analyses, as described in Leonardi et al. (2023) \doi{10.1111/ecog.06481}. The package includes datasets of palaeoclimate reconstructions, present observations, and future projections from multiple climate models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EvolEcolGroup/pastclim}
  \item \url{https://evolecolgroup.github.io/pastclim/}
  \item Report bugs at \url{https://github.com/EvolEcolGroup/pastclim/issues}
}

}
\author{
\strong{Maintainer}: Andrea Manica \email{am315@cam.ac.uk} [copyright holder]

Authors:
\itemize{
  \item Michela Leonardi
}

Other contributors:
\itemize{
  \item Emily Y. Hallet [contributor]
  \item Robert Beyer [contributor]
  \item Mario Krapp [contributor]
  \item Andrea V. Pozzi [contributor]
}

}
\keyword{internal}
