% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastclim.R
\docType{package}
\name{pastclim}
\alias{pastclim}
\title{pastclim}
\description{
This `R` library is designed to provide an easy way to extract and manipulate palaeoclimate
reconstructions for ecological and anthropological analyses.
}
\details{
The functionalities of `pastclim` are described in 
 [Leonardi et al. (2023)](https://doi.org/10.1111/ecog.06481). Please cite it if you
use `pastclim` in your research.
}
