% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_series.R
\name{region_series}
\alias{region_series}
\title{Extract a time series of climate variables for a region}
\usage{
region_series(time_bp, bio_variables, dataset, path_to_nc = NULL)
}
\arguments{
\item{time_bp}{time slices in years before present (negative). The slices needs
to exist in the dataset. To check which slices are available, you can use
\code{get_time_steps}.}

\item{bio_variables}{vector of names of variables to be extracted}

\item{dataset}{string defining the dataset to use. If set to "custom",
then a single nc file is used from "path_to_nc"}

\item{path_to_nc}{the path to the custom nc file containing the paleoclimate
reconstructions. All the variables of interest need to be included in this file.

location_slice
location_series
region_slice
region_series
slice_region_series}
}
\description{
This function extracts a time series of one or more climate variables for a given
dataset covering a region (or the whole world). The function returns a
SpatRasterDataset \code{terra::sds} object, with
each variable as a sub-dataset.
}
