% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_region_series.R
\name{sample_rs_fixed}
\alias{sample_rs_fixed}
\title{Sample the same locations from a region time series}
\usage{
sample_rs_fixed(x, size, method = "random", replace = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{a \code{terra::SpatRasterDataset} returned
by \code{region_series}}

\item{size}{number of points sampled; the same
locations across all time steps}

\item{method}{one of the sampling methods from \code{terra::spatSample}. It
defaults to "random"}

\item{replace}{boolean determining whether we sample with replacement}

\item{na.rm}{boolean determining whether NAs are removed}
}
\value{
a data.frame with the sampled cells and their respective values for
the climate variables.
}
\description{
Internal function for fixed sampling from \code{sample_region_series}, 
used when a single size is given.
}
\keyword{internal}
