% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bp_to_index.R
\name{time_bp_to_index}
\alias{time_bp_to_index}
\title{Find the closest index to a given time in years BP}
\usage{
time_bp_to_index(time_bp, time_steps)
}
\arguments{
\item{time_bp}{vector of times BP}

\item{time_steps}{time steps for which reconstructions are available}
}
\value{
the indeces for the relevant time steps
}
\description{
Internal function
}
\keyword{internal}
