% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bp_to_index.R
\name{time_bp_to_index}
\alias{time_bp_to_index}
\title{Convert a time BP to indeces in a netcdf file.}
\usage{
time_bp_to_index(time_bp, path_to_nc)
}
\arguments{
\item{time_bp}{vector of times BP}

\item{path_to_nc}{path to nc file}
}
\description{
Internal function
}
