% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_for_location.R
\name{time_series_for_locations}
\alias{time_series_for_locations}
\title{Extract a time series of bioclimatic variables for one or more locations.}
\usage{
time_series_for_locations(x, bio_variables, dataset, path_to_nc = NULL)
}
\arguments{
\item{x}{a 2 column matrix (with columns `longitude`, ranging
-180 to 180, and `latitude`, from -90 to 90), or a vector of cell numbers.}

\item{bio_variables}{vector of names of variables to be extracted.}

\item{dataset}{string defining the dataset to use.}

\item{path_to_nc}{the path to the directory containing the downloaded resonstructions.
Leave it unset if you are using the companion `pastclimData` to store datasets.}
}
\description{
This function extract a time series of local climate for
 a set of locations
}
