% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_docs.R
\name{WorldClim_2.1}
\alias{WorldClim_2.1}
\alias{WorldClim_2.1_10m}
\alias{WorldClim_2.1_5m}
\title{Documentation for the WorldClim datasets}
\description{
The WorldClim version 2.1 climate datasets are based on a mean 1970-2000,
and were released in January 2020. In \code{pastclim}, the datasets are given
a date of 1985 CE (the mid-point of the period of interest), corresponding to
a time_bp of 35.
}
\details{
If you use this dataset, make sure to cite the original publication:

Fick, S.E. and R.J. Hijmans, 2017. WorldClim 2: new 1km spatial resolution
climate surfaces for global land areas. International Journal of Climatology 37 (12): 4302-4315.
\doi{doi.org/10.1002/joc.5086}

There are monthly climate data for minimum, mean, and maximum temperature, precipitation,
and 19 “bioclimatic” variables.

There are multiple resolutions of
10 arc-minutes, 5 arc-minutes, 2.5 arc-minute and 0.5 arc-minutes. The resolution
of interest can be obtained by changing the ending of the dataset name
"WorldClim_2.1_XXXm", e.g. WorldClim_2.1_10m or WorldClim_2.1_5m
Currently, only 10m and 5m are available on \code{pastclim}.
}
